+++ common/src/main/java/com/alien/Alien.java
@@ -7,7 +7,6 @@ import com.alien.common.registry.init.AlienEntityTypes;
 import com.alien.common.registry.init.AlienGameEvents;
 import com.alien.common.registry.init.AlienInfections;
 import com.alien.common.registry.init.AlienItems;
-import com.alien.common.registry.init.AlienLifecycles;
 import com.alien.common.registry.init.block.AlienChitinBlocks;
 import com.alien.common.registry.init.block.AlienResinBlocks;
 import com.alien.common.registry.init.item.AlienArmorItems;
@@ -43,7 +42,6 @@ public class Alien {
 
         // Custom
         AlienInfections.initialize();
-        AlienLifecycles.initialize();
 
         // Functionality
         AlienCompostingChances.initialize();