+++ common/src/main/java/com/alien/common/model/lifecycle/growth/GrowthStageKey.java
@@ -2,7 +2,7 @@ package com.alien.common.model.lifecycle.growth;
 
 import net.minecraft.world.entity.EntityType;
 
-public record AlienGrowthStageKey(
+public record GrowthStageKey(
     EntityType<?> host,
     EntityType<?> other
 ) {}