+++ fabric/src/main/java/com/avp/fabric/data/growth_stages/GrowthConstants.java
@@ -1,0 +1,16 @@
+package com.avp.fabric.data.growth_stages;
+
+import java.util.concurrent.TimeUnit;
+
+public class GrowthConstants {
+
+    public static final int CHESTBURSTER_GROWTH_TIME_IN_TICKS = (int) TimeUnit.MINUTES.toSeconds(5) * 20;
+
+    public static final int DRONE_GROWTH_TIME_IN_TICKS = (int) TimeUnit.MINUTES.toSeconds(10) * 20;
+
+    public static final int WARRIOR_GROWTH_TIME_IN_TICKS = (int) TimeUnit.MINUTES.toSeconds(15) * 20;
+
+    public static final int PRAETORIAN_GROWTH_TIME_IN_TICKS = (int) TimeUnit.MINUTES.toSeconds(20) * 20;
+
+    public static final int ROYAL_CHESTBURSTER_GROWTH_TIME_IN_TICKS = (int) TimeUnit.MINUTES.toSeconds(5) * 20;
+}