+++ neoforge/src/main/java/com/avp/neoforge/service/NeoForgeRegistryService.java
@@ -1,6 +1,5 @@
 package com.avp.neoforge.service;
 
-import com.alien.common.model.lifecycle.AlienLifecycle;
 import com.alien.common.model.lifecycle.infection.AlienInfection;
 import com.bvanseg.just.functional.tuple.Tuple2;
 import com.bvanseg.just.functional.tuple.Tuple3;
@@ -10,6 +9,7 @@ import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
+import net.minecraft.server.packs.resources.PreparableReloadListener;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.Mob;
@@ -48,8 +48,6 @@ public class NeoForgeRegistryService implements RegistryService {
 
     private final List<Supplier<? extends AlienInfection<?, ?>>> alienInfectionSuppliers;
 
-    private final List<Supplier<AlienLifecycle>> alienLifecycleSuppliers;
-
     private final List<Supplier<? extends Item>> azureLibItemIdentitySuppliers;
 
     private final List<Tuple4<Supplier<? extends ItemLike>, Float, Boolean, Boolean>> compostableData;
@@ -63,6 +61,8 @@ public class NeoForgeRegistryService implements RegistryService {
     private final List<LiteralArgumentBuilder<CommandSourceStack>> literalArgumentBuilders;
 
     private final List<NetworkHandler<?>> networkHandlers;
+
+    private final List<PreparableReloadListener> reloadListeners;
 
     private final List<Tuple3<Supplier<VillagerProfession>, Integer, List<VillagerTrades.ItemListing>>> villagerTradeData;
 
@@ -88,7 +88,6 @@ public class NeoForgeRegistryService implements RegistryService {
         ).collect(Collectors.toMap(Function.identity(), NeoForgeRegistryService::createDeferredRegistry));
 
         this.alienInfectionSuppliers = new ArrayList<>();
-        this.alienLifecycleSuppliers = new ArrayList<>();
         this.azureLibItemIdentitySuppliers = new ArrayList<>();
         this.compostableData = new ArrayList<>();
         this.entityAttributeSupplierPairs = new ArrayList<>();
@@ -96,6 +95,7 @@ public class NeoForgeRegistryService implements RegistryService {
         this.furnaceFuelPairs = new ArrayList<>();
         this.literalArgumentBuilders = new ArrayList<>();
         this.networkHandlers = new ArrayList<>();
+        this.reloadListeners = new ArrayList<>();
         this.villagerTradeData = new ArrayList<>();
     }
 
@@ -122,12 +122,6 @@ public class NeoForgeRegistryService implements RegistryService {
     ) {
         alienInfectionSuppliers.add(alienInfectionSupplier);
         return alienInfectionSupplier;
-    }
-
-    @Override
-    public Supplier<AlienLifecycle> registerAlienLifecycle(Supplier<AlienLifecycle> alienLifecycleSupplier) {
-        alienLifecycleSuppliers.add(alienLifecycleSupplier);
-        return alienLifecycleSupplier;
     }
 
     @Override
@@ -174,6 +168,12 @@ public class NeoForgeRegistryService implements RegistryService {
     }
 
     @Override
+    public PreparableReloadListener registerReloadListener(String id, PreparableReloadListener listener) {
+        reloadListeners.add(listener);
+        return listener;
+    }
+
+    @Override
     public void registerVillagerTrade(
         Supplier<VillagerProfession> villagerProfessionSupplier,
         int level,
@@ -192,10 +192,6 @@ public class NeoForgeRegistryService implements RegistryService {
 
     public List<Supplier<? extends AlienInfection<?, ?>>> getAlienInfectionSuppliers() {
         return alienInfectionSuppliers;
-    }
-
-    public List<Supplier<AlienLifecycle>> getAlienLifecycleSuppliers() {
-        return alienLifecycleSuppliers;
     }
 
     public List<Supplier<? extends Item>> getAzureLibItemIdentitySuppliers() {
@@ -224,6 +220,10 @@ public class NeoForgeRegistryService implements RegistryService {
 
     public List<NetworkHandler<?>> getNetworkHandlers() {
         return networkHandlers;
+    }
+
+    public List<PreparableReloadListener> getReloadListeners() {
+        return reloadListeners;
     }
 
     public List<Tuple3<Supplier<VillagerProfession>, Integer, List<VillagerTrades.ItemListing>>> getVillagerTradeData() {