+++ CHANGELOG.md
@@ -23,7 +23,15 @@
 
 ## 🛠 Data Pack
 - Added `#avp:runner_hosts` entity tag.
+- Added `#avp:runners` entity tag.
+- Added `#avp:prowlers` entity tag.
+- Added `#avp:crushers` entity tag.
 - Updated `#avp:hosts` to be composed of `#avp:runner_hosts` entity tag.
+- Updated `#avp:xenomorphs` tag to include new runner alien line entity tags.
+- Updated hive layer entity tags to include new runner alien line entity tags.
+- Updated variant entity tags to include new runner alien entity types.
 
 ## 🔬 Technical Changes
-- Upgraded alien lifecycle registry to support entity tags.
+- Upgraded alien lifecycle registry to support host entity tags.
+- Added full datagen support for alien growth stages. Alien lifecycle registry has been removed.
+- Alien lifecycle host entity tags have been migrated to individual alien growth stages.